Clear views array on job_create event
Fixes juttle/outrigger#83

When running a different program from our original program, there
was an issue with the old views getting reinstantiated because they
still existed in redux store. This change resets the views in the store
to an empty object so new views are only instantiated when we have info
about them.