input.getValues: change to return a promise
When a user updates a value in an input, we make a call to `/prepare` to get a fresh state of input values and definitions and then update the redux store with that result. This means that if a consumer calls inputs.getValues before our `/prepare` returns, they won't get the most update set of input values.

To fix this, getValues return a promise instead of an immediate value. To keep track of whether we are in the process of updating the value (i.e. making a `/prepare` call and stuff), add BEGIN_UPDATE_INPUT_VALUE and END_UPDATE_INPUT_VALUE actions and maintain a `updatingValueState` property in the store. getValues waits for `updatingValueState === 'COMPLETED'` if it isn't already before resolving the promise.