Refactor juttle-service client api
Add error reporting for all requests. Responses that are not 200s should
get throw as an error.

Wrap what-wg fetch so it returns a bluebird promise instead of an es6
promise. Much better.

Split up functions in api so they can be imported as a standalone (not
having to instantiate a class).

Rename from OutriggerAPI to JuttleServiceHttp

Add tests for juttle-service api.