job-manager: job-status events
When the readyState of the internal WebSocket changes, fire job-status
event. Hook this event to the view class so users can subscribe to it as
well. Event returns the JobStatus code of the job (0= CONNECTING,
1= OPEN, etc)

Arose from a need in juttle-viewer to keep track of the status of a
running job. Now we have it.