Make JobStatus different from websocket readyState
Previously we were using webSocket readyState for JobManager status. The
two are not exactly the same. For instance- a job is technically
starting when the post request to create the job has been made, and not
just when the websocket connection is initializing.

Instead use own internal statuses: 'STOPPED', 'RUNNING', 'STARTING'.