Close websocket on job_end event
Was experiencing this issue when trying to implement a better run button
on juttle-viewer. Currently we await the server to issue a close event,
and as a result was reporting the job as running, when it really had
recieved a `job_end` event.

It's better for the job-manager to close itself (and therefore
get set to `STOPPED` state).